fix: error while remove a used domain

+++ app/Http/Controllers/DomainController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Domain;
+use App\Models\User;
 use Illuminate\Http\Request;
 
 class DomainController extends Controller
@@ -26,6 +27,10 @@ class DomainController extends Controller
     }
     public function destroy(Domain $domain)
     {
+        $existedUserWithDomain = User::where('domain_id', $domain->id)->get()->first();
+        if($existedUserWithDomain) {
+            return back()->with('error', 'Vous ne pouvez pas supprimer un domaine déjà utilisé');
+        }
         $domain->delete();
         return redirect()->route('domain.create')->with('success', 'Domaine supprimé avec succès.');
     }
